move widget to @ts/core/r1
