+++ packages/devextreme/js/__internal/core/r1/base_props.ts
