+++ packages/devextreme/js/__internal/pagination/common/base_pagination_props.ts
@@ -1,9 +1,9 @@
 import type { DisplayMode } from '@js/common';
 import type { PageSize } from '@js/ui/pagination_types';
+import { BaseWidgetDefaultProps, type BaseWidgetProps } from '@ts/core/r1/base_props';
 
 import messageLocalization from '../../../localization/message';
 import type { EventCallback } from '../../core/r1/event_callback';
-import { BaseWidgetDefaultProps, type BaseWidgetProps } from '../base_props';
 
 export interface BasePaginationProps extends BaseWidgetProps {
   isGridCompatibilityMode?: boolean;
